requirements/testing.txt
-r base.txt

coverage==3.7.1
django-debug-toolbar==0.11.0
django-discover-runner==1.0
